chore: rename to langchain_callback submodule
