--- src/greptimeai/__init__.py
@@ -1,2 +1,0 @@
-# ruff: noqa: F401
-from greptimeai.langchain_tracker.callback import GreptimeCallbackHandler
