+++ tests/langchain_callback/test_langchain_helper_func.py
@@ -1,4 +1,4 @@
-from greptimeai.langchain_tracker import (
+from greptimeai.langchain_callback import (
     _get_serialized_id,
     _get_serialized_streaming,
     _get_user_id,
